refactor(panel): rename files and folder for panel
Fixes #21
