--- projects/ngx-bulma/src/lib/panel/bulma-panel-block/bulma-panel-block.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'bu-panel-block',
-  templateUrl: './bulma-panel-block.component.html',
-  styleUrls: ['./bulma-panel-block.component.css']
-})
-export class BulmaPanelBlockComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
