--- projects/ngx-bulma/src/lib/panel/bulma-panel-heading/bulma-panel-heading.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'bu-panel-heading',
-  templateUrl: './bulma-panel-heading.component.html',
-  styleUrls: ['./bulma-panel-heading.component.css']
-})
-export class BulmaPanelHeadingComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
