--- projects/ngx-bulma/src/lib/panel/bulma-panel-tabs/bulma-panel-tabs.component.html
@@ -1,3 +1,0 @@
-<p class="panel-tabs">
-    <ng-content></ng-content>
-  </p>
