--- projects/ngx-bulma/src/lib/panel/bulma-panel-tabs/bulma-panel-tabs.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'bu-panel-tabs',
-  templateUrl: './bulma-panel-tabs.component.html',
-  styleUrls: ['./bulma-panel-tabs.component.css']
-})
-export class BulmaPanelTabsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
