+++ projects/ngx-bulma/src/lib/panel/panel-block/panel-block.component.html
