+++ projects/ngx-bulma/src/lib/panel/panel-block/panel-block.component.ts
@@ -1,0 +1,17 @@
+import { Component, OnInit,  ViewEncapsulation, ChangeDetectionStrategy  } from '@angular/core';
+
+@Component({
+  selector: 'bu-panel-block',
+  templateUrl: './panel-block.component.html',
+  styleUrls: ['./panel-block.component.css'],
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class BulmaPanelBlockComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
