+++ projects/ngx-bulma/src/lib/panel/panel-heading/panel-heading.component.html
