+++ projects/ngx-bulma/src/lib/panel/panel-heading/panel-heading.component.ts
@@ -1,0 +1,17 @@
+import { Component, OnInit , ViewEncapsulation, ChangeDetectionStrategy  } from '@angular/core';
+
+@Component({
+  selector: 'bu-panel-heading',
+  templateUrl: './panel-heading.component.html',
+  styleUrls: ['./panel-heading.component.css'],
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class BulmaPanelHeadingComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
