+++ projects/ngx-bulma/src/lib/panel/panel-tabs/panel-tabs.component.html
@@ -1,0 +1,3 @@
+<p class="panel-tabs">
+  <ng-content></ng-content>
+</p>
