+++ projects/ngx-bulma/src/lib/panel/panel-tabs/panel-tabs.component.ts
@@ -1,0 +1,17 @@
+import { Component, OnInit, ViewEncapsulation, ChangeDetectionStrategy } from '@angular/core';
+
+@Component({
+  selector: 'bu-panel-tabs',
+  templateUrl: './panel-tabs.component.html',
+  styleUrls: ['./panel-tabs.component.css'],
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class BulmaPanelTabsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
